chore(restructure): split editor -> editing/ai_filters + editing/prompts + utils/filters
Phase 1 step 5: separate the Gemini-driven filter generator. The
VideoEditor class moves to openshorts/editing/ai_filters.py; the two
long Gemini prompt strings move to openshorts/editing/prompts.py as
functions; the previously private filter helpers (sanitize_filter_string,
enforce_zoompan_output_size, split_filter_chain) move to
openshorts/utils/filters.py so the future motion-graphics and audio
compositors can reuse them.

VideoEditor still re-exposes the helpers as static/classmethods so the
existing characterization tests pass unchanged. Shim at editor.py keeps
`from editor import VideoEditor` working. Tests stay 62/62 green.
